% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequentItemsets.R
\name{FrequentItemsets}
\alias{FrequentItemsets}
\title{A function to find frequent itemsets using the wonderful spmf library}
\usage{
FrequentItemsets(x, algo, minsup, clean = T)
}
\arguments{
\item{x}{is the output of the fnction df2SPMFBasket function. It must contain a toSendSPMF storing a tibble with a basket column. This basket is sent to spmf java library.}

\item{algo}{is the name of the algorithm to be used. The possible values are : "Apriori","Apriori_TID","FPGrowth_itemsets","Eclat","HMine","FIN","LCMFreq","AprioriClose","Charm_bitset","dCharm_bitset","LCM","FPMax","Charm_MFI","defME". Please check SPMF documentation at http://www.philippe-fournier-viger.com/spmf/index.php?link=documentation.php}

\item{minsup}{is the value of the minimum support for an itemset to be seen as frequent. to get all the itemset present in more than 40 percent of the transactions, type either "40\%" or "0.4"}

\item{clean}{if clean=F the text files sent to spmf will be kept in the working directory.}
}
\value{
a dataframe with three columns. itemset contains all the frequent itemsets, that is a list of products. support is the number of times this itemsets occurs, and frequence is support divided by the total number of transactions
}
\description{
This function sends a list of baskets to spmf and gets back the most frequent itemsets with their support and frequence.
}
\examples{




}
