% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OriginalIDs.R
\name{OriginalIDs}
\alias{OriginalIDs}
\title{A function to match original items form their numeric value (when recoded in the spmf format)}
\usage{
OriginalIDs(x, evLev)
}
\arguments{
\item{x}{is a set of IDs supporting a given sequence}

\item{evLev}{the original IDs list}
}
\description{
This function is not to be used by the user but to be called by FrequentSequences.
}
\examples{




}
