% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OriginalItems.R
\name{OriginalItems}
\alias{OriginalItems}
\title{A function to match original items form their numeric value (when recoded in the spmf format)}
\usage{
OriginalItems(x, evLev)
}
\arguments{
\item{x}{is a basket or a sequence in the sense of the SPMF format. It is an element from the basket slot of the output of a df2SPMFBasket function call. I}

\item{evLev}{the original factor levels to recode to}
}
\description{
This function is not to be used by the user but to be called by FrequentItemsets or more likely by FrequentSequences.
}
\examples{




}
\keyword{IDs}
