% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequentialRules.R
\name{SequentialRules}
\alias{SequentialRules}
\title{A function to find frequential rules using the wonderful spmf library}
\usage{
SequentialRules(x, algo, minsup = 0.5, minconf = 0.6, clean = F,
  windowSize = "")
}
\arguments{
\item{x}{is the output of the function df2SPMFSequence function. It must contains a toSendSPMF storing a tibble with a basket column. This basket is sent to spmf java library.}

\item{algo}{is the name of the algorithm to be used. The possible values are : Please check SPMF documentation at http://www.philippe-fournier-viger.com/spmf/index.php?link=documentation.php}

\item{minsup}{is the value of the minimum support for a sequence to be seen as frequent. To get all the sequences present in more than 40 percent of the database sequences, type either "40\%" or "0.4"}

\item{clean}{if clean=F the text files sent to and received from spmf will be kept in the working directory.}

\item{windowSize}{the TRuleGrowth searches for rules X==>Y within a given window size, based on the sequences who contains all the items of X before all the items of Y whithin the windowsize.}

\item{Those}{time parameters are only considered when the input is a timed sequence. The only algorithme to be used are then "Fournier08-Closed+time" and "HirateYamana"}
}
\value{
a dataframe with three columns. sequence contains all the frequent sequences. support is the number of times this sequence occurs, and frequence is support divided by the total number of sequences
}
\description{
This function sends a list of sequences to spmf and gets back Sequential Rules sequences with their support and confidence.
}
\examples{




}
