% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToBasket.R
\name{ToBasket}
\alias{ToBasket}
\title{Basket creating function}
\usage{
ToBasket(x)
}
\arguments{
\item{x}{a dataframe which must contains variable ID, itemset, eventNum, and possibly time}
}
\description{
This function is not to be used directly, it's called by df2spmfBasket
}
\examples{
ToBasket()
}
\keyword{baskets}
