% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToSequence.R
\name{ToSequence}
\alias{ToSequence}
\title{Sequence creating function}
\usage{
ToSequence(x)
}
\arguments{
\item{x}{a dataframe which must contains variable ID, itemset, eventNum, and possibly time}
}
\description{
This function is not to be used directly, it's called by df2spmfSequence
}
\examples{
ToBasket()
}
\keyword{baskets}
