% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToTimedBasket.R
\name{ToTimedBasket}
\alias{ToTimedBasket}
\title{Basket creating function, taking care of time}
\usage{
ToTimedBasket(x)
}
\arguments{
\item{x}{a dataframe which must contains variable ID, itemset, eventNum and time}
}
\description{
This function is not to be used directly, it is called by df2spmfBasket
}
\examples{
ToTimedBasket()
}
\keyword{baskets}
