% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransactionBuilder.R
\name{TransactionBuilder}
\alias{TransactionBuilder}
\title{A function to create sequences ID according to the delay between two events}
\usage{
TransactionBuilder(df, ID, event, time, timeFormat, timeMax = "",
  timeMin = 0, timeUnit = "secs", parallel = F)
}
\arguments{
\item{df}{is a dataframe}

\item{ID}{is the ID, for instance of a customer. Within this ID the function will create one or several sequences, according to time between events.}

\item{time}{is the time of the event}

\item{timeFormat}{is the format in which the time is encoded}

\item{timeMax}{if the time elapsed between two events is more than timeMax, a new sequence is started}

\item{timeMin}{if the time elapsed between two events is less than timeMin, these events are considered to happen together : they belong to the same itemset}

\item{parallel}{if True the function will use the several cores available to split the building of the sequence. It appears to be useful above 1 million of lines, depending, of course, of the specs of the computer.}
}
\value{
a dataframe with three columns. sequence contains all the frequent sequences. support is the number of times this sequence occurs, and frequence is support divided by the total number of sequences
}
\description{
Input is a data frame and affects a sequence ID to the events related to an ID, built on the time elapsed between an event and thi previous event.
}
\examples{

}
\keyword{Transaction}
\keyword{database}
